openshift/s3-backup/docker/pyproject.toml
[tool.poetry]
name = "s3-backup"
version = "0.1.0"
description = "Prune backup images"
authors = ["Sybrand Strauss <sybrand.strauss@gov.bc.ca>"]
license = "Apache License, Version 2.0, January 2004"

[tool.poetry.dependencies]
python = "^3.6"
aiobotocore = "^1.4.2"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
